apitradex/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==5.0.0
pymysql==1.1.2
python-dotenv==1.0.0